sema: move var_init typechecking to separate file
